requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.1
defusedxml==0.6.0
dj-rest-auth==1.1.1
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
idna==2.10
isort==5.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.12.9
py3dns==3.2.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
validate-email==1.3
wrapt==1.12.1
